Move test/* to test/units/ and Add test/utils
