+++ package.json
@@ -16,7 +16,7 @@
     "lint": "node_modules/tslint/bin/tslint -c ./tslint.json --project . **/*.ts",
     "lint-fix": "node_modules/tslint/bin/tslint -c ./tslint.json --project . --fix **/*.ts",
     "webpack": "serverless webpack",
-    "test": "./node_modules/.bin/mocha -r ts-node/register src/test/**/*.spec.ts src/test/**/**/*.spec.ts",
+    "test": "./node_modules/.bin/mocha -r ts-node/register src/test/units/**/*.spec.ts src/test/units/**/**/*.spec.ts",
     "test-all": "npm-run-all lint webpack test"
   },
   "dependencies": {
